requirements.txt
aniso8601==9.0.1
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
clickclick==20.10.2
colorama==0.4.4
connexion==2.7.0
decorator==5.0.7
defusedxml==0.7.1
ecdsa==0.14.1
entrypoints==0.3
Flask==1.1.2
Flask-Cors==3.0.8
flask-marshmallow==0.14.0
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
idna==2.10
importlib-metadata==4.0.1
inflection==0.5.1
ipykernel==5.5.3
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
MarkupSafe==1.1.1
marshmallow==3.11.1
marshmallow-sqlalchemy==0.24.3
mistune==0.8.4
mysqlclient==2.0.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.19.5
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.0
packaging==20.9
pandas==1.0.4
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
plotly==4.14.3
prometheus-client==0.10.1
prompt-toolkit==3.0.18
pyasn1==0.4.8
pycparser==2.20
Pygments==2.8.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-jose==3.2.0
pytz==2021.1
pywin32==300
pywinpty==1.0.0
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
requests==2.25.1
retrying==1.3.3
rsa==4.7.2
Send2Trash==1.5.0
six==1.15.0
SQLAlchemy==1.4.11
terminado==0.9.4
testpath==0.4.4
tornado==6.1
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
zipp==3.4.1
connexion[swagger-ui]